.olds/iot_brain/pyproject.toml
[project]
name = "iot-brain"
version = "0.1.0"
description = ""
readme = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv==1.0.1",
]
